sreality/requirements.txt
flask
psycopg2-binary
scrapy
